requirements.txt
requests~=2.27.1
pycryptodome
Pillow~=9.0.0
onnxruntime~=1.9.0
numpy~=1.21.2